refactor : admin test
